requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.13.0
certifi==2026.2.25
charset-normalizer==3.4.7
click==8.3.2
distro==1.9.0
ecdsa==0.19.2
fastapi==0.135.3
groq==0.37.1
grpcio==1.80.0
h11==0.16.0
h2==4.3.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
hyperframe==6.1.0
idna==3.11
jsonpatch==1.33
jsonpointer==3.1.1
langchain-core==1.2.30
langchain-groq==1.1.2
langsmith==0.7.32
numpy==2.4.4
orjson==3.11.8
packaging==26.1
portalocker==3.2.0
protobuf==7.34.1
psycopg2-binary==2.9.12
pyasn1==0.6.3
pydantic==2.13.1
pydantic_core==2.46.1
python-dotenv==1.2.2
python-jose==3.5.0
python-multipart==0.0.26
PyYAML==6.0.3
qdrant-client==1.17.1
requests==2.33.1
requests-toolbelt==1.0.0
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.49
starlette==1.0.0
tenacity==9.1.4
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uuid_utils==0.14.1
uvicorn==0.44.0
xxhash==3.6.0
zstandard==0.25.0
